+++ e2e/typescript/src/assets/fixture.ts
@@ -134,8 +134,8 @@ assertType({
   boxSizing: 'outline-box',
 });
 assertType({
-  // @ts-expect-error "1" is invalid value for "overflow"
-  overflow: '1',
+  // @ts-expect-error "1" is invalid value for "overflowX"
+  overflowX: '1',
 });
 assertType({
   // @ts-expect-error "paddingLeft" cannot be numeric value
@@ -188,13 +188,31 @@ assertType({
   color: 'var(--customColor)',
 });
 
+assertType({
+  '@media screen and (max-width: 992px)': {
+    ...typedMixin,
+  },
+  // @ts-expect-error FIXME - this is supposed to pass!
+  '@media(forced-colors: active)': {
+    ...typedMixin,
+    color: 'var(--customColor)',
+  },
+  ':after': {
+    ...typedMixin,
+  },
+  ':before': {
+    ...typedMixin,
+    color: 'var(--customColor)',
+  },
+});
+
 // Strict selectors
 //
 
 assertType({
   ':hover': {
-    // @ts-expect-error "1" is invalid value for "overflow"
-    overflow: '1',
+    // @ts-expect-error "1" is invalid value for "overflowX"
+    overflowX: '1',
     // @ts-expect-error "padding" is banned
     padding: 0,
     // @ts-expect-error "paddingLeft" cannot be numeric value
@@ -213,21 +231,45 @@ assertType({
 //
 
 assertType({
+  // @ts-expect-error "1" is invalid value for "overflowX"
   ':hover:focus': {
-    // @ts-expect-error "1" is invalid value for "overflow"
-    overflow: '1',
-    // @ts-expect-error "padding" is banned
+    overflowX: '1',
+  },
+});
+
+assertType({
+  // @ts-expect-error "padding" is banned
+  ':hover:focus': {
     padding: 0,
-    // @ts-expect-error "paddingLeft" cannot be numeric value
+  },
+});
+
+assertType({
+  // @ts-expect-error "paddingLeft" cannot be numeric value
+  ':hover:focus': {
     paddingLeft: 5,
   },
 });
+
+assertType({
+  // @ts-expect-error "1" is invalid value for "overflowX", padding is banned, paddingLeft cannot be a numeric value
+  ':hover:focus': {
+    overflowX: 'scroll',
+    padding: 0,
+    paddingLeft: 5,
+  },
+});
+
 assertType({
   ':hover:focus': {
     // @ts-expect-error Object is not assignable to CSS property
     zIndex: { color: 'red' },
+    opacity: { color: 'red' }, // < no error here, TS only reports the first error in this case
+  },
+  ':hover:active': {
     // @ts-expect-error Object is not assignable to CSS property
     opacity: { color: 'red' },
+    zIndex: { color: 'red' }, // < no error here, TS only reports the first error in this case
   },
 });
 
@@ -235,37 +277,33 @@ assertType({
 //
 
 assertType({
+  // @ts-expect-error "1" is invalid value for "overflowX", padding is banned, paddingLeft cannot be a numeric value
   '.foo': {
     '.baz': {
-      // @ts-expect-error "1" is invalid value for "overflow"
-      overflow: '1',
-      // @ts-expect-error "padding" is banned
+      overflowX: '1',
       padding: 0,
-      // @ts-expect-error "paddingLeft" cannot be numeric value
       paddingLeft: 5,
     },
   },
 });
 assertType({
+  // @ts-expect-error outline-box is an invalid value for box-sizing
   '.foo': {
-    // @ts-expect-error outline-box is an invalid value for box-sizing
     boxSizing: 'outline-box',
 
     '.bar': {
-      // @ts-expect-error outline-box is an invalid value for box-sizing
-      boxSizing: 'outline-box',
+      boxSizing: 'outline-box', // < no error here, TS only reports the error for the whole object
     },
   },
 });
 assertType({
+  // @ts-expect-error outline-box is an invalid value for box-sizing
   '.foo': {
-    // @ts-expect-error type check still fails on outline-box, not on any other line
     boxSizing: 'outline-box',
     zIndex: 1,
 
     '.bar': {
-      // @ts-expect-error type check still fails on outline-box, not on any other line
-      boxSizing: 'outline-box',
+      boxSizing: 'outline-box', // < no error here, TS only reports the error for the whole object
       zIndex: 1,
     },
   },
@@ -274,14 +312,38 @@ assertType({
   '.foo': {
     // @ts-expect-error Object is not assignable to CSS property
     zIndex: { color: 'red' },
-    // @ts-expect-error Object is not assignable to CSS property
-    opacity: { color: 'red' },
+    opacity: { color: 'red' }, // < no error here, TS only reports the first error in this case
 
     '.bar': {
-      // @ts-expect-error Object is not assignable to CSS property
-      zIndex: { color: 'red' },
-      // @ts-expect-error Object is not assignable to CSS property
-      opacity: { color: 'red' },
+      zIndex: { color: 'red' }, // < no error here, TS only reports the first error in this case
+      opacity: { color: 'red' }, // < no error here, TS only reports the first error in this case
     },
+  },
+});
+
+// Fallback values
+assertType({
+  color: ['red', 'blue'],
+  paddingLeft: [0, '2px'],
+  zIndex: [10, 20],
+  ':hover': {
+    color: ['red', 'blue'],
+  },
+  ':hover:active': {
+    zIndex: [2],
+    color: ['red', 'blue'],
+    paddingLeft: [0, '2px'],
+  },
+});
+
+assertType({
+  paddingLeft: [
+    // @ts-expect-error "paddingLeft" cannot be numeric value
+    2,
+    '2px',
+  ],
+  // @ts-expect-error "paddingLeft" cannot be numeric value
+  ':hover:active': {
+    paddingLeft: ['2px', 2],
   },
 });