+++ packages/core/src/constants.ts
@@ -11,6 +11,15 @@ export const SEQUENCE_HASH_LENGTH = 7;
 export const SEQUENCE_PREFIX = '___';
 
 /** @internal */
+export const DEBUG_SEQUENCE_SEPARATOR = '_';
+
+/** @internal */
+export const SEQUENCE_SIZE =
+  process.env.NODE_ENV === 'production'
+    ? SEQUENCE_PREFIX.length + SEQUENCE_HASH_LENGTH
+    : SEQUENCE_PREFIX.length + SEQUENCE_HASH_LENGTH + DEBUG_SEQUENCE_SEPARATOR.length + SEQUENCE_HASH_LENGTH;
+
+/** @internal */
 export const DEFINITION_LOOKUP_TABLE: Record<SequenceHash, LookupItem> = {};
 
 // indexes for values in LookupItem tuple