+++ packages/core/src/runtime/compileCSS.test.ts
@@ -38,6 +38,20 @@ describe('compileCSS', () => {
     `);
   });
 
+  it('handles array of values', () => {
+    expect(
+      compileCSS({
+        ...defaultOptions,
+        property: 'color',
+        value: ['red', 'blue'],
+      }),
+    ).toMatchInlineSnapshot(`
+      Array [
+        ".foo{color:red;color:blue;}",
+      ]
+    `);
+  });
+
   it('handles at-rules', () => {
     expect(
       compileCSS({
@@ -100,6 +114,23 @@ describe('compileCSS', () => {
       Array [
         ".foo:before{padding-left:10px;}",
         ".rtl-foo:before{padding-right:10px;}",
+      ]
+    `);
+  });
+
+  it('handles rtl properties with fallback values', () => {
+    expect(
+      compileCSS({
+        ...defaultOptions,
+        property: 'paddingLeft',
+        value: [0, '10px'],
+        rtlProperty: 'paddingRight',
+        rtlValue: [0, '10px'],
+      }),
+    ).toMatchInlineSnapshot(`
+      Array [
+        ".foo{padding-left:0;padding-left:10px;}",
+        ".rtl-foo{padding-right:0;padding-right:10px;}",
       ]
     `);
   });