+++ packages/core/src/shorthands/border.test.ts
@@ -39,4 +39,21 @@ describe('border', () => {
       borderTopColor: 'red',
     });
   });
+
+  it('for fallback value arrays', () => {
+    expect(border(['2px', '4px'], ['solid', 'dashed'], ['red', 'green'])).toEqual({
+      borderBottomColor: ['red', 'green'],
+      borderBottomStyle: ['solid', 'dashed'],
+      borderBottomWidth: ['2px', '4px'],
+      borderLeftColor: ['red', 'green'],
+      borderLeftStyle: ['solid', 'dashed'],
+      borderLeftWidth: ['2px', '4px'],
+      borderRightColor: ['red', 'green'],
+      borderRightStyle: ['solid', 'dashed'],
+      borderRightWidth: ['2px', '4px'],
+      borderTopColor: ['red', 'green'],
+      borderTopStyle: ['solid', 'dashed'],
+      borderTopWidth: ['2px', '4px'],
+    });
+  });
 });