+++ packages/devtools/jest.config.js
@@ -1,0 +1,9 @@
+module.exports = {
+  displayName: 'devtools',
+  preset: '../../jest.preset.js',
+  transform: {
+    '^.+\\.[tj]sx?$': 'babel-jest',
+  },
+  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
+  coverageDirectory: '../../coverage/packages/devtools',
+};