+++ packages/devtools/public/devtools-page.js
@@ -1,0 +1,7 @@
+chrome.devtools.panels.elements.createSidebarPane('💅 Griffel', sidebarPanel => {
+  sidebarPanel.setPage('index.html');
+
+  sidebarPanel.onShown.addListener(devtoolsWindow => {
+    devtoolsWindow.postMessage({ visible: true }, '*');
+  });
+});