+++ packages/webpack-loader/src/transformSync.ts
@@ -22,7 +22,7 @@ export function transformSync(sourceCode: string, options: TransformOptions): Tr
   // Parse the code first so Babel will use user's babel config for parsing
   // During transforms we don't want to use user's config
   const babelAST = Babel.parseSync(sourceCode, {
-    caller: { name: 'make-styles' },
+    caller: { name: 'griffel' },
 
     filename: options.filename,
     inputSourceMap: options.inputSourceMap,