+++ packages/babel-preset/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - @griffel/babel-preset
 
-This log was last generated on Wed, 09 Mar 2022 13:56:06 GMT and should not be manually modified.
+This log was last generated on Wed, 06 Apr 2022 13:28:28 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 1.2.0
+
+Wed, 06 Apr 2022 13:28:28 GMT
+
+### Minor changes
+
+- feat: export shaker evaluator (olfedias@microsoft.com)
+- Bump @griffel/core to v1.2.0
 
 ## 1.1.1